Added 2 rendering features
- Glow (Glow size, color, etc)
Makes an entity glow... who wouldve guessed. Size does not scale up with camera zoom and opacity is currently unadjustable.
- Upgrade Color
Changes/Overides the color of the panel behind the tank in the class tree and upgrade menu. I plan to have a "Default upgrade color" config that defaults to the current coloring system.